created component to edit channel
